fix(compaction): 修复数值转换类型安全问题
- 修复了测试代码中的整数转换溢出风险
- 将 std::stoull 结果显式转换为 uint64_t 类型
- 更新 TODO 清单状态标记，反映并发策略实现进度
- 完成 compaction 入口锁细化设计文档更新

+++ tests/compaction_test.cpp
@@ -56,7 +56,7 @@ uint64_t GetMaxFileNumberOnDisk(const std::string& dir) {
                      [](unsigned char c) { return std::isdigit(c); })) {
       continue;
     }
-    max_id = std::max(max_id, std::stoull(stem));
+    max_id = std::max(max_id, static_cast<uint64_t>(std::stoull(stem)));
   }
   return max_id;
 }
